pyproject.toml
[tool.poetry]
name = "repl_python3_EV-9D9"
version = "0.1.0"
description = ""
authors = ["Buddy Cross budsta95@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
alpaca-trade-api = "^0.51.0"
discord = "^1.0.1"
python-dotenv = "^0.14.0"
replit = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
